+++ lib/formotion/row_type/slider_row.rb
@@ -1,3 +1,5 @@
+motion_require 'base'
+
 module Formotion
   module RowType
     class SliderRow < Base
@@ -6,7 +8,7 @@ module Formotion
       SLIDER_VIEW_TAG = 1200
 
       def build_cell(cell)
-        cell.selectionStyle = UITableViewCellSelectionStyleNone
+        cell.selectionStyle = self.row.selection_style || UITableViewCellSelectionStyleNone
         slideView = UISlider.alloc.initWithFrame(CGRectZero)
         cell.accessoryView = cell.editingAccessoryView = slideView
         row.range ||= (1..10)