+++ lib/formotion/row_type/string_row.rb
@@ -1,3 +1,5 @@
+motion_require 'base'
+
 module Formotion
   module RowType
     class StringRow < Base
@@ -19,6 +21,7 @@ module Formotion
       # Also does the layoutSubviews swizzle trick
       # to size the UITextField so it won't bump into the titleLabel.
       def build_cell(cell)
+        cell.selectionStyle = self.row.selection_style || UITableViewCellSelectionStyleBlue
         field = UITextField.alloc.initWithFrame(CGRectZero)
         field.tag = TEXT_FIELD_TAG
 
@@ -40,6 +43,7 @@ module Formotion
         field.autocorrectionType = row.auto_correction if row.auto_correction
         field.clearButtonMode = row.clear_button || UITextFieldViewModeWhileEditing
         field.enabled = row.editable?
+        field.inputAccessoryView = input_accessory_view(row.input_accessory) if row.input_accessory
 
         add_callbacks(field)
 
@@ -69,6 +73,7 @@ module Formotion
           end
         end
 
+        field.font = BW::Font.new(row.font) if row.font
         field.placeholder = row.placeholder
         field.text = row_value
         cell.addSubview(field)
@@ -128,13 +133,7 @@ module Formotion
       end
 
       def on_select(tableView, tableViewDelegate)
-        super or _on_select(tableView, tableViewDelegate)
-      end
-
-      def _on_select(tableView, tableViewDelegate)
-        if !row.editable?
-          return
-        else
+        if row.editable?
           row.text_field.becomeFirstResponder
         end
       end
@@ -143,6 +142,43 @@ module Formotion
       def update_text_field(new_value)
         self.row.text_field.text = row_value
       end
+
+      # Creates the inputAccessoryView to show
+      # if input_accessory property is set on row.
+      # :done is currently the only supported option.
+      def input_accessory_view(input_accessory)
+        case input_accessory
+        when :done
+          @input_accessory ||= begin
+            tool_bar = UIToolbar.alloc.initWithFrame([[0, 0], [0, 44]])
+            tool_bar.autoresizingMask = UIViewAutoresizingFlexibleWidth
+            tool_bar.barStyle = UIBarStyleBlack
+            tool_bar.translucent = true
+
+            left_space = UIBarButtonItem.alloc.initWithBarButtonSystemItem(
+                UIBarButtonSystemItemFlexibleSpace,
+                target: nil,
+                action: nil)
+
+            done_button = UIBarButtonItem.alloc.initWithBarButtonSystemItem(
+                UIBarButtonSystemItemDone,
+                target: self,
+                action: :done_editing)
+
+            tool_bar.items = [left_space, done_button]
+
+            tool_bar
+          end
+        else
+          nil
+        end
+      end
+
+      # Callback for "Done" button in input_accessory_view
+      def done_editing
+        self.row.text_field.endEditing(true)
+      end
+
     end
   end
 end