+++ lib/formotion/row_type/subform_row.rb
@@ -1,3 +1,5 @@
+motion_require 'base'
+
 module Formotion
   module RowType
     class SubformRow < Base
@@ -5,6 +7,7 @@ module Formotion
       LABEL_TAG=1001
 
       def build_cell(cell)
+        cell.selectionStyle = self.row.selection_style || UITableViewCellSelectionStyleBlue
         cell.accessoryType = cell.editingAccessoryType = UITableViewCellAccessoryDisclosureIndicator
 
         cell.contentView.addSubview(self.display_key_label)
@@ -25,6 +28,7 @@ module Formotion
         end
 
         display_key_label.highlightedTextColor = cell.textLabel.highlightedTextColor
+        nil
       end
 
       def update_cell(cell)