+++ lib/formotion/row_type/switch_row.rb
@@ -1,10 +1,12 @@
+motion_require 'base'
+
 module Formotion
   module RowType
     class SwitchRow < Base
       include BW::KVO
 
       def build_cell(cell)
-        cell.selectionStyle = UITableViewCellSelectionStyleNone
+        cell.selectionStyle = self.row.selection_style || UITableViewCellSelectionStyleNone
         switchView = UISwitch.alloc.initWithFrame(CGRectZero)
         switchView.accessibilityLabel = (row.title || "") + " Switch"
         cell.accessoryView = cell.editingAccessoryView = switchView