+++ CHANGELOG.md
@@ -1,8 +1,24 @@
-## 1.3 - ??
+## 1.3.1 - April 15, 2013
+
+### Features
+
+- Added `:input_accessory` property, which currently accepts `:done` as the value. This will add a `UIToolbar` above the keyboard with a "Done" button the user can tap to dismiss the keyboard.
+
+- Added `Row#on_delete` callback which occurs when a row is swipe-to-delete'd.
+
+### Bug Fixes
+
+-  `:image` rows which are not editable do not show the "plus" icon.
+
+- Fixed bugs in `:template` rows (see [#100](https://github.com/clayallsopp/formotion/issues/100) and [#101](https://github.com/clayallsopp/formotion/pull/101)).
+
+## 1.3 - March 25, 2013
 
 ### Features
 
 - Added `:currency` row type, which automatically presents an entered number into the current locale's format (i.e. "$4,003.45" or "€ 3.004,35")
+
+- Added `on_delete` callback for when a `Row` is deleted
 
 - Correctly handle `on_tap` for `Row` objects, regardless of whether or not they are `:button`s.
 
@@ -11,6 +27,8 @@
 - Fixed a crash that occured when rapidly serializing a form.
 
 - Template rows are now persisted correctly.
+
+- `ImageRow`s which are disabled will not show the "plus" icon
 
 ## 1.2 - January 3, 2013